started refactor, browser helper module
