+++ test/testHelpers.js
@@ -56,3 +56,5 @@ function messagePresent (text) {
 	return null;
 }
 
+
+
